Move ArgumentHandler class to a Libs #86
