+++ libs/CMakeLists.txt
@@ -3,3 +3,4 @@ add_subdirectory(ECS)
 add_subdirectory(Error)
 add_subdirectory(Transisthor)
 add_subdirectory(GameComponents)
+add_subdirectory(ArgumentHandler)
